Verify review + adapt some parts
